django/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
environs==9.5.0
requests==2.31.0
